Rename models/shoes/dtos to models/shoes/dto
